Added two functions to control the robot's movement.
The "turn_right" function turns the robot 90 degrees to the right, while the "jump" function makes the robot jump over an obstacle by turning left, moving forward one square, turning right, moving forward one square, turning right
again, and finally moving forward one square.

The "while" loop checks if the robot has reached the goal. If not, it checks if there is a wall in front of the robot.
If there is a wall, the "jump" function is called to make the robot jump over the obstacle. Otherwise, the robot moves forward one square.